fix: s/auth/authAddress
