+++ packages/auth-client/src/actions/app/createChannel.ts
@@ -1,6 +1,7 @@
 import { AsyncUnwrapped, unwrap } from "../../errors";
 import { Client } from "../../clients/createClient";
 import { HttpResponse, post } from "../../clients/transports/http";
+import { AuthMethod } from "../../types";
 
 export type CreateChannelArgs = CreateChannelRequest;
 export type CreateChannelResponse = AsyncUnwrapped<HttpResponse<CreateChannelAPIResponse>>;
@@ -13,7 +14,7 @@ interface CreateChannelRequest {
   expirationTime?: string;
   requestId?: string;
   redirectUrl?: string;
-  acceptMethods?: string[];
+  acceptMethods?: AuthMethod[];
 }
 
 export interface CreateChannelAPIResponse {
