+++ packages/auth-client/src/actions/app/status.test.ts
@@ -24,7 +24,7 @@ describe("status", () => {
       ip: "127.0.0.1",
       userAgent: "Mozilla/5.0",
     },
-    acceptMethods: ["custody", "auth"],
+    acceptMethods: ["custody", "authAddress"],
   };
 
   test("constructs API request", async () => {
