+++ packages/auth-client/src/index.ts
@@ -3,3 +3,4 @@ export { generateNonce } from "siwe";
 export * from "./errors";
 export * from "./actions";
 export * from "./clients";
+export * from "./types";
