+++ packages/auth-client/src/types.ts
@@ -1,0 +1,10 @@
+import type { SiweMessage } from "siwe";
+
+export type AuthMethod = "custody" | "authAddress";
+
+export type FarcasterResourceParams = {
+  fid: number;
+  method?: AuthMethod;
+};
+
+export type SignInMessageParams = Partial<SiweMessage> & FarcasterResourceParams;
